// src/types/IOpenSys.ts
export default interface IOpenSys {
  phoneNumber?: string;
  businessId?: string; // Y-tunnus
  company?: string;
  firstName?: string;
  lastName?: string;
  address?: string;
  city?: string;
  email?: string;
  website?: string;
  used?: Date | string;
  result?: string;
  resultDate?: Date | string;
  comments?: string;
  title?: string;
  industry?: string;
  staffClass?: string;
  revenueClass?: string;
  postalCode?: string;
  mobileNumber?: string;
  importMessage?: string;
  companyLanguage?: string;
  createdAt?: Date;
  updatedAt?: Date;
}